fix onnx2tf inputs for input/output quant dtype

+++ ultralytics/engine/exporter.py
@@ -1014,7 +1014,8 @@ class Exporter:
             output_integer_quantized_tflite=self.args.int8,
             quant_type="per-tensor",  # "per-tensor" (faster) or "per-channel" (slower but more accurate)
             custom_input_op_name_np_data_path=np_data,
-            input_output_quant_dtype=io_quant_dtype,
+            input_quant_dtype=io_quant_dtype,
+            output_quant_dtype=io_quant_dtype,
             enable_batchmatmul_unfold=True,  # fix lower no. of detected objects on GPU delegate
             output_signaturedefs=True,  # fix error with Attention block group convolution
             optimization_for_gpu_delegate=True,
